If the DOM changes textfield focus, make the IME work properly.
Fix for http://b/issue?id=2219166

Requires a change to external/webkit

Remove the old change to update the WebTextView when a key is pressed,
since the IME does not generate key events.  Instead, when the focus
changes, and the IME is serving the WebTextView, immediately clear
the cursor and update the WebTextView, so the user can continue
typing.