bugfix endfling case (bug 2300534)
When the menu comes up over the activity we call removeCallbacks(mFlingRunnable);
The effect of that is that our mTouchMode variable is left dangling in the fling
state.  This change calls endFling() after removeCallbacks.  In the specific bug
being addressed we depend on reportScrollStateChange() because we avoid doing
requeries while scrolling.