Fix incorrect calculation of start and end indices of changed text
- The problem is taht the calculation of indices does not take the
  difference in original and modified text lengths into account.
- If reportExtractedText() is called from the onDraw() function,
  mInputMethodState is not reset.

The error can occur when the user is editing formatted text using a full screen editor.
One example of formatted text is phone numbers (e.g. dashes are added to North American
numbers by the PhoneNumberFormattingTextWatcher class).
If the two first digits are removed from the number 555-666-777, the new number becomes 566--777.
This is due incorrect calculation of start and end indices of the changed text. The error is in
the TextView class and the problem is that the calculation does not take the difference in original
and modified text lengths into account.

Change-Id: If12d0f571873a6f1874102377432affb64ee21e8