Focus the IME on WebView when no textfield is in focus.
This is to fix a particular case where the navigation cache
does not recognize that a textfield is now in focus.  A message
was sent from the WebCore thread telling us to open the IME on
the textfield.  Open it with the WebView as the served view.

Fix for http://b/issue?id=2457459