Adds to DumpRenderTree the ability to look for Android-specific results
These Android-specific results will be added to
external/webkit/LayoutTests/platform/android-<js-engine> and are used in
preference to the generic expected results.

The JavaScript engine to use is read from the JS_ENGINE environment variable
used by the build system or can be overridden on the command line. If neither
is set, it defaults to JSC.

Change-Id: Ia8d107ced3968a5c061fd6f0f57451419bab6b27